1. made some naming changes according to the  conventions
