--- src/main/java/com/cs544/dao/ILocationDao.java
@@ -1,11 +1,0 @@
-package com.cs544.dao;
-
-import com.cs544.domain.Location;
-
-import java.util.List;
-
-public interface ILocationDao {
-
-    boolean save(Location location);
-    List<Location> findAll();
-}
