+++ src/main/java/com/cs544/dao/LocationRepository.java
@@ -2,7 +2,8 @@ package com.cs544.dao;
 
 import com.cs544.domain.Location;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
-
+@Repository
 public interface LocationRepository extends CrudRepository<Location,Long> {
 }
