+++ src/main/java/com/cs544/domain/Admin.java
@@ -7,7 +7,7 @@ import javax.persistence.Table;
 import java.util.List;
 @Entity
 @Table(name = "adminstrator")
-public class Admin extends IUser{
+public class Admin extends User {
 	
 	public Admin() {}
